+++ packages/core/src/dom/media/simple-hls/index.ts
@@ -1,4 +1,4 @@
 import { SpfMediaMixin } from '@videojs/spf/dom';
-import { HTMLVideoElementHost } from '../video-host';
+import { HTMLVideoElementHost } from '../host';
 
 export class SimpleHlsMedia extends SpfMediaMixin(HTMLVideoElementHost) {}
